+++ packages/sync-server/app/tasks/VdsNcSignerEndOfWorkingPeriodChecker.js
@@ -3,23 +3,23 @@ import { ScheduledTask } from 'shared/tasks';
 import { log } from 'shared/services/logging';
 import { Op, Sequelize } from 'sequelize';
 
-export class VdsNcSignerExpiryChecker extends ScheduledTask {
+export class VdsNcSignerEndOfWorkingPeriodChecker extends ScheduledTask {
   constructor(context) {
-    const conf = config.schedules.vds.signerExpiryChecker;
+    const conf = config.schedules.vds.signerEndOfWorkingPeriodChecker;
     super(conf.schedule, log);
     this.config = conf;
     this.context = context;
   }
 
   getName() {
-    return 'VdsNcSignerExpiryChecker';
+    return 'VdsNcSignerEndOfWorkingPeriodChecker';
   }
 
   async run() {
     const { VdsNcSigner } = this.context.store.models;
     const expired = await VdsNcSigner.findAll({
       where: {
-        notAfter: { [Op.lt]: Sequelize.NOW },
+        workingPeriodEnd: { [Op.lt]: Sequelize.NOW },
       },
       // hard limit: there really should only ever be zero or one
       // expired signers at any given time, but just in case:
@@ -30,7 +30,7 @@ export class VdsNcSignerExpiryChecker extends ScheduledTask {
       return Promise.resolve();
     }
 
-    log.info(`${expired.length} VDS-NC signer(s) expired`);
+    log.info(`${expired.length} VDS-NC signer(s) have reached the end of their working period`);
 
     return Promise.all(
       expired.map(async signer => {