+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ### Tweaks ⚖️
 
 - Tweaked localisation to prepare for loading it from the db
+- Scheduled task `VdsNcSignerExpiryChecker` renamed to `VdsNcSignerEndOfWorkingPeriodChecker`, `schedules` config key has changed in consequence.
 
 ### Bug fixes 🐛